requirements.txt
flask
pycryptodome
numpy
opencv-python 